remove skyimager cli
